pyproject.toml
[tool.poetry]
name = "workedon"
version = "0.1.1"
description = "CLI tool for daily work logging"
authors = ["viseshrp <viseshrprasad@gmail.com>"]
readme = "README.rst"

[tool.poetry.scripts]
workedon = "workedon.__main__:main"

[tool.poetry.dependencies]
python = "^3.6.2"
click = "^8.0.3"
dateparser = "^1.1.0"
tzlocal = "^3.0"
"backports.zoneinfo" = [
    { version = "^0.2.1", python = "<3.9" }
]
click-default-group = "^1.2.2"
peewee = "^3.14.8"

[tool.poetry.dev-dependencies]
pytest = "*"
flake8 = "*"
pyinvoke = "*"
safety = "*"
pylint = [
    { version = "*", python = "~=3.6" }
]
ipdb = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
